Add files via upload
Latest version of hack-princeton hack: collab with Mark Seda, Liz Strobel, and Aric Lu.
Takes signal from Rasberri Pi to know when to look for images on Drop Box.
Process Image to determine if there is a misplaced item, and if so, identifies the item and uses text-to-speech to alert people of the item it found. Uses IBM Watson.